Software License Changes
The Creative Commons Attribution-ShareAlike 4.0 International
("CC BY-SA 4.0") license originally used with the php-mvc software and
source code is not valid for software licensing. Since the original
license is not valid for software licensing, the copyright holder is now
upgrading the license from its original version.

As of August 24, 2016, the php-mvc software will be licensed under the
GNU Affero General Public License v3 ("AGPLv3").

All previously distributed copies and modifications of this software are
hereby licensed under the AGPLv3 license retroactively.